# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  # TO DO: CHANGE THE DESCRIPTION
  description: |-
    This OpenAPI document describes the application of a software which allows people to stay in touch with other users of the system.
    With the inspiration of `WhatsApp`, this app goals is to create an
    open-source one with less security in it
  version: "1"
  contact:
    name: Diego Antonio Guzman Aguirre
    email: guzmanaguirre.2006301@studenti.uniroma1.it
    url: https://github.com/guizzyy
servers:
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '201':
          description: User creation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': {$ref: '#/components/responses/BadRequest'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /conversations:
    get:
      tags: ["user profile"]
      summary: get all conversations of the user
      description: |
        Retrieve all the conversations the user is in
        both with another user and groups
      operationId: getMyConversations
      responses:
        '200':
          description: chats loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Homepage'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}

    post:
      tags: ["user action"]
      summary: start a conversation with a user
      description: |
        With a given user, you can start a new conversation with him
        and this will be saved in your home along with the others already
        started
      operationId: startConversation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Conversation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /group:
    post:
      tags: ["user action"]
      summary: create a new group
      description: |
        Create a new group with a name and an optional photo,
        the founder can add whoever he likes from all the users in the
        application
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '201':
          description: group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
  /conversations/{convID}/open:
    parameters:
      - $ref: '#/components/parameters/ConversationID'

    get:
      tags: ["chats"]
      summary: get a conversation with a user or group
      description: |
        Retrieve a conversation with a user or a group
      operationId: getConversation
      parameters:
        - name: lastID
          in: query
          required: true
          description: last message ID
          schema:
            type: integer
            example: 100
            readOnly: true
            description: id message
      responses:
        '200':
          description: chat loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List-of-Message'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /memberships/{convID}:
    parameters:
      - $ref: '#/components/parameters/ConversationID'

    get:
      tags: ["chats"]
      summary: get a list of conversation members
      description: |
        Ask for a list of all the members in the conversation,
        if private the list will have length 2
      operationId: getMembers
      responses:
        '200':
          description: list of members loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List-of-User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: ["groups"]
      summary: add a user to a group
      description: |
        Add a member to a group. It can leave anytime
      operationId: addToGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: user added correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
...
  /conversations/{convID}/messages/{messID}/reactions:
    parameters:
      - $ref: '#/components/parameters/ConversationID'
      - $ref: '#/components/parameters/MessageID'

    get:
      tags: ["chats"]
      summary: get a list of comments
      description: |
        Get a list of all the comments for an existed message,
        it could include the sender also and the list is made of
        user and emoji sent
      operationId: getComments
      responses:
        '200':
          description: comments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List-of-Reaction'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError'}


    put:
      tags: ["chats"]
      summary: react to a message
      description: |
        React to a message sent in a conversation,
        a user can only react one emoji per message
      operationId: commentMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reaction'
      responses:
        '201':
          description: message correctly reacted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}

    delete:
      tags: ["chats"]
      summary: delete a reaction of a message
      description: |
        Delete a reaction from a message you have put before
      operationId: uncommentMessage
      responses:
        '204':
          description: reaction deleted correctly
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalServerError'}
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      description: authentication code mode
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
    Unauthorized:
      description: The access token is missing or it's expired
    NotFound:
      description: Resource not found in the database
  parameters:
...
    ConversationID:
      schema:
        type: integer
        example: 2
      name: convID
      in: path
      required: true
      description: ID of the conversation
    MessageID:
      schema:
        type: integer
        example: 3
      name: messID
      in: path
      required: true
      description: ID of the message
  schemas:
    User:
      description: User schema
      type: object
      properties:
        id:
          type: integer
          example: 10
          readOnly: true
          description: user ID
        username:
          type: string
          example: Diego
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: username
        photo:
          $ref: '#/components/schemas/Photo'
    Group:
      description: group schema
      type: object
      properties:
        id:
          type: integer
          example: 11
          readOnly: true
          description: group ID
        name:
          type: string
          example: Class
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: group name
        photo:
          $ref: '#/components/schemas/Photo'
    Message:
      description: message schema
      type: object
      properties:
        id:
          type: integer
          example: 12
          readOnly: true
          description: message ID
        text:
          $ref: '#/components/schemas/Text'
        timestamp:
          type: string
          format: date-time
          minLength: 1
          maxLength: 999
          description: date and hour of the message
        sender:
          $ref: '#/components/schemas/User'
        photo:
          $ref: '#/components/schemas/Photo'
        status:
          type: string
          enum:
            - "Unreceived"
            - "Received"
            - "Read"
          description: status of the message
        reactions:
          $ref: '#/components/schemas/List-of-Reaction'
    Conversation:
      description: conversation schema
      type: object
      properties:
        messages:
          $ref: '#/components/schemas/List-of-Message'
    Homepage:
      description: homepage schema
      type: object
      properties:
        convs:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
          minItems: 0
          maxItems: 999
          description: conversations started
    Reaction:
      description: reaction schema
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        messageId:
          type: integer
          example: 1234
          description: id of the message reacted
          readOnly: true
        emoji:
          type: string
          example: "🤣"
          minLength: 1
          maxLength: 4
          pattern: "^[\\u1F600-\\u1F64F\\u1F300-\\u1F5FF\\u1F680-\\u1F6FF\\u1F700-\\u1F77F\\u1F780-\\u1F7FF\\u1F800-\\u1F8FF\\u1F900-\\u1F9FF\\u1FA00-\\u1FA6F\\u1FA70-\\u1FAFF\\u2600-\\u26FF\\u2700-\\u27BF\\u2300-\\u23FF\\u2B50\\u23F0\\u231A\\u25AA-\\u25FE\\u2B06\\u2194-\\u2199\\u21A9-\\u21AA\\u2753-\\u2755\\u274C\\u274E]$"
          description: single reacted emoji
    List-of-Message:
      description: List of message in a conversation
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          minItems: 0
          maxItems: 999
          description: messages found
    List-of-Reaction:
      description: List of reaction schema
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'
          minItems: 0
          maxItems: 999
          description: reactions found
    List-of-User:
      description: List of users schema
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
          minItems: 0
          maxItems: 999
          description: users found
    Photo:
      description: Photo schema
      type: object
      properties:
        file:
          type: string
          format: binary
          minLength: 1
          maxLength: 999
          description: file of the photo
    Text:
      description: message text
      type: object
      properties:
        text:
          type: string
          example: This is a message!
          pattern: '^.*?$'
          minLength: 1
          maxLength: 255
          description: the text of the message
    Notification:
      description: notification sent to the user
      type: object
      properties:
        report:
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 250
          example: "Resource updated successfully"
          description: notify user of the operation result