# service/api/doLogin.go
package api

import (
	"encoding/json"
	"git.guizzyy.it/WASAText/service/api/reqcontext"
	"git.guizzyy.it/WASAText/service/utilities"
	"github.com/julienschmidt/httprouter"
	"net/http"
)

func (rt *_router) doLogin(w http.ResponseWriter, r *http.Request, params httprouter.Params, context reqcontext.RequestContext) {
	// Get the username for log in the request body
	var userLog utilities.User
	if err := json.NewDecoder(r.Body).Decode(&userLog); err != nil {
		context.Logger.WithError(err).Error("json login decode error")
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	pUser := &userLog

	// Check if the username provided has the correct format
	if check, err := rt.checkStringFormat(userLog.Username); err != nil {
		context.Logger.WithError(err).Error("error during string format check")
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	} else if !check {
		context.Logger.Error(utilities.ErrUsernameString)
		http.Error(w, utilities.ErrUsernameString.Error(), http.StatusBadRequest)
		return
	}

	// Ask the database if it is a new/existing user and get their ID and photo
	isNew, err := rt.db.LogUser(pUser)
	if err != nil {
		context.Logger.WithError(err).Error("error during logUser db")
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	response := utilities.LoginResponse{}

	if !isNew {
		// The user already exists
		w.WriteHeader(http.StatusOK)
		if userLog.Photo, err = rt.GetFile(userLog.Photo); err != nil {
			context.Logger.WithError(err).Error("error during GetFile")
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		response.Message = "Login successful"
		response.UserLogged = userLog
	} else {
		// The user is new
		if err = rt.CreateUserDir(userLog.ID, context, w); err != nil {
			context.Logger.WithError(err).Error("error during create user dir")
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		w.WriteHeader(http.StatusCreated)
		response.Message = "User created successfully"
		response.UserLogged = userLog
	}

	w.Header().Set("Content-Type", "application/json")
	if err = json.NewEncoder(w).Encode(response); err != nil {
		context.Logger.WithError(err).Error("json login encode error")
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
}